code/src/Backend/requirements.txt
flask==3.0.2
pymongo==4.6.1
pdf2image==1.17.0
Pillow==10.2.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.48
python-dotenv==1.0.0
scikit-learn==1.3.0